Refactored type casting

+++ PixelGameEngineTinker/PixelGameEngineTinker/Settings.cpp
@@ -3,4 +3,4 @@
 
 const std::string Settings::Assets::SPRITES_FILE_PATH = "./sprites.db";
 
-const float Settings::WorldChunk::TILE_RENDER_MIN_CELL_SIZE = 1.0f; // [ hard code ]
+const long double Settings::WorldChunk::TILE_RENDER_MIN_CELL_SIZE = 1.0f; // [ hard code ]
